Kill channel thread when connection is closed during callback. #69
Exceptions that occur within a subscriber callback are caught.
When a connection is closed, expectedly or unexpectedly, an exception is
thrown onto each channel thread to stop them. However, if that exception
is thrown while a subscription callback is executing, it gets caught,
the channel thread is never stopped and the channel exception handlers
are never run.

This fix wraps exceptions that should kill a channel thread and makes
sure they don't get caught.