AMQP: change fromURI to activate TLS if amqps
Previously, use of `amqps://` changed default port, but didn't activate
TLS. Change `Network.AMQP.fromURI` to activate TLS using `TLSTrusted` if
`amqps://` protocol is used. This is likely the assumption anyway, since
it was surprising to find that only the port was changed.

This is a breaking change, although I'm unsure whether it should be
considered a bugfix or a new feature. Old behaviour is easily retained
by changing `amqps://u:p@h` to `amqp://u:p@h:5671`.